Merge duplicate for-loops into single for loop.
This change also fixes an issue that arises when the number of elements
in the array is odd. Previously, when the array length is odd, the second
for loop attempts to access invalid elements (e.g. ary[1.5]), which caused
half the elements to not be rendered. Since the merged for loop no longer
divides the array length by 2, this change also fixes the odd-array-length
bug.