Rotate bars and ensure all bars are displayed.
These changes cause the bars to be drawn from the bottom up rather
than the top down. Additionally, extra code was added to the random
number generation phase to ensure that 0 never appears in the random
numbers array, which ensures that there will never be a bar with 0
height in the graph.